+++ lib/marv/project.rb
@@ -1,5 +1,4 @@
 require 'pathname'
-require 'compass'
 
 module Marv
   class Project
@@ -92,13 +91,8 @@ module Marv
       if File.exists?(self.config_file)
         config.merge!(load_ruby_config(self.config_file))
       else
-        # Old format of config file
-        if File.exists?(File.join(self.root, 'config.json'))
-          config.merge!(convert_old_config)
-        else
-          raise Error, "Could not find the config file, are you sure you're in a
-          marv project directory?"
-        end
+        raise Error, "Could not find the config file, are you sure you're in a
+        marv project directory?"
       end
 
       @config = config
@@ -113,36 +107,6 @@ module Marv
     end
 
     private
-
-    def convert_old_config
-      require 'json'
-
-      # Let the user know what is going to happen
-      @task.say("It looks like you are using the old JSON-format config. Marv will now try converting your config to the new Ruby format.")
-      @task.ask(" Press any key to continue...")
-
-      begin
-        old_file_name = File.join(self.root, 'config.json')
-        # Parse the old config format, convert keys to symbols
-        @config = JSON.parse(File.open(old_file_name).read).inject({}){|memo,(k,v)| memo[k.to_sym] = v; memo}
-
-        @task.create_file(@config_file) do
-          # Find the config.tt template, and parse it using ERB
-          config_template_path = @task.find_in_source_paths(File.join(['config', 'config.tt']))
-          parse_erb(File.expand_path(config_template_path))
-        end
-      rescue Exception => e
-        @task.say "Error while building new config file:", Thor::Shell::Color::RED
-        @task.say e.message, Thor::Shell::Color::RED
-        @task.say "You'll need to either fix the error and try again, or manually convert your config.json file to Ruby format (config.rb)"
-        exit
-      end
-
-      @task.say "Success! Double-check that all your config values were moved over, and you can now delete config.json.", Thor::Shell::Color::GREEN
-
-      # We now have a Ruby config file, so we can continue loading as normal
-      return load_ruby_config(self.config_file)
-    end
 
     def load_ruby_config(file)
       config = {}